// Nepnep/src/types/types.ts
declare global {

    interface EmscriptenModule {
        preRun: any[]
        postRun: any[]
        print: (text: string) => void
        printErr: (text: string) => void
        canvas: HTMLElement | null
        setStatus: (text: string) => void
        monitorRunDependencies: () => void
        preinitializedWebGPUDevice?: GPUDevice
    }

    interface Window {
        Module?: EmscriptenModule
    }

    interface GPUDevice {
    }

    interface Navigator {
        gpu?: any
    }
    
}
export enum LayerType {
    Rectangle,
    Ellipse,
    Path,
    Text
}

export enum CanvasMode {
    None,
    Dragging,
    Inserting,
    Pencil,
    Resizing,
    Translating,
    SelectionNet,
    Pressing,
    RightClick,
}

export enum Side {
    Top= 1,
    Bottom = 2,
    Left = 4,
    Right = 8
}

export type Color = {
    r: number
    g: number
    b: number
}

export type Camera = {
    x: number
    y: number
    zoom: number
}

export type RectangleLayer = {
    type: LayerType.Rectangle
    x: number
    y: number
    height: number
    width: number
    fill: Color
    stroke: Color
    opacity: number
    connerRadius: number
}

export type EllipseLayer = {
    type: LayerType.Ellipse
    x: number
    y: number
    height: number
    width: number
    fill: Color
    stroke: Color
    opacity: number
}

export type PathLayer = {
    type: LayerType.Path
    x: number
    y: number
    height: number
    width: number
    fill: Color
    stroke: Color
    opacity: number
    points: number[][]
}

export type TextLayer = {
    type: LayerType.Text
    x: number
    y: number
    height: number
    width: number
    fill: Color
    stroke: Color
    opacity: number
    text: string
    fontSize: number
    fontWeight: number
    fontFamily: string
}

export type Layer = RectangleLayer | EllipseLayer | PathLayer | TextLayer

export type Point = {
    x: number
    y: number
}

export type XYWH = {
    x: number
    y: number
    width: number
    height: number
}

export type CanvasState =
    | { mode: CanvasMode.None }
    | { mode: CanvasMode.Dragging; origin: Point | null }
    | { mode: CanvasMode.Inserting; layerType: LayerType }
    | { mode: CanvasMode.Pencil }
    | { mode: CanvasMode.Resizing; initialBounds: XYWH; corner: Side }
    | { mode: CanvasMode.Translating; current: Point }
    | { mode: CanvasMode.SelectionNet; origin: Point; current?: Point }
    | { mode: CanvasMode.Pressing; origin: Point }
    | { mode: CanvasMode.RightClick }
    